+++ R/rposterior.R
@@ -27,7 +27,7 @@
 #'     \item {"os"} {A numeric matrix or data frame. Each row should contain
 #'       the largest order statistics for a block of data.  These needs not
 #'       be ordered because they are sorted inside \code{rpost}. If a block
-#'       contains fewer than \code{dim(as.matrix(data)[2]} order statistics
+#'       contains fewer than \code{dim(as.matrix(data))[2]} order statistics
 #'       then the corresponding row should be padded by \code{NA}s. If
 #'       \code{ros} is supplied then only the largest \code{ros} values in
 #'       each row are used.}
@@ -39,9 +39,9 @@
 #'   be supplied when \code{model = "pp"}.  If \code{model = "gp"} and
 #'   \code{thresh} is not supplied then \code{thresh = 0} is used and
 #'   \code{data} should contain threshold excesses.
-#' @param noy A numeric scalar  The number of observations per block, where a
-#'   block is often a year.  Only relevant, and must be supplied, if
-#'   \code{model = "pp"}.
+#' @param noy A numeric scalar. The number of blocks of observations,
+#'   excluding any missing values.  A block is often a year.
+#'   Only relevant, and must be supplied, \code{model = "pp"}.
 #' @param use_noy A logical scalar.  Only relevant if model is "pp".  By
 #'   default (\code{use_noy = FALSE}) sampling is based on a likelihood in
 #'   which the number of blocks (years) is set equal to the number of threshold
@@ -106,7 +106,17 @@
 #'
 #' @return An object (list) of class \code{"evpost"}, which has the same
 #'   structure as an object of class "ru" returned from \code{\link[rust]{ru}}.
-#'   If \code{model == "bingp"} then this list contains in addition
+#'   In addition this list contains
+#'   \itemize{
+#'     \item{\code{model}:} The argument \code{model} to \code{rpost}
+#'       detailed above.
+#'     \item{\code{data}:} The argument \code{data} to \code{rpost}
+#'       detailed above.
+#'   }
+#'   If \code{model == "pp"} then this list also contains the argument
+#'     \code{noy} to \code{rpost} detailed above.
+#'
+#'   If \code{model == "bingp"} then this list also contains
 #'   \itemize{
 #'     \item{\code{bin_sim_vals}:} {An \code{n} by 1 numeric matrix of values
 #'       simulated from the posterior for the binomial
@@ -173,6 +183,7 @@ rpost <- function(n, model = c("gev", "gp", "bingp", "pp", "os"), data, prior,
                   init_ests = NULL, mult = 2, use_phi_map = FALSE, ...) {
   #
   model <- match.arg(model)
+  save_model <- model
   # Check that the prior is compatible with the model.
   # If an evdbayes prior has been set make the "model" attribute of the
   # prior equal to "gev"
@@ -414,6 +425,11 @@ rpost <- function(n, model = c("gev", "gp", "bingp", "pp", "os"), data, prior,
       temp$bin_logf_args <- temp_bin$bin_logf_args
     }
     class(temp) <- "evpost"
+    temp$model <- save_model
+    temp$data <- data
+    if (save_model == "pp") {
+      temp$noy <- noy
+    }
     return(temp)
   }
   #
@@ -515,6 +531,11 @@ rpost <- function(n, model = c("gev", "gp", "bingp", "pp", "os"), data, prior,
     temp$bin_logf_args <- temp_bin$bin_logf_args
   }
   class(temp) <- "evpost"
+  temp$model <- save_model
+  temp$data <- data
+  if (save_model == "pp") {
+    temp$noy <- noy
+  }
   return(temp)
 }